raaa vercel.... i'm starting to think it's not right for the task, but anyway i added a workaround to make it not store any files if the fi;e system is read only

+++ app.py
@@ -42,7 +42,16 @@ with open("config.json", "r") as the_file:
 app = flask.Flask(__name__, static_folder="assets")
 #driver = webdriver.Chrome()
 remote_zip = RemoteZip(ZIP_FILE_LINK)
-local_images = os.listdir("assets/fetched")
+readonly_fs = False
+try:
+    local_images = os.listdir("assets/fetched")
+except FileNotFoundError:
+    try:
+        os.mkdir("assets/fetched")
+        local_images = os.listdir("assets/fetched")
+    except OSError:
+        readonly_fs = True
+        local_images = []
 raw_remote_images = remote_zip.filelist
 remote_images = []
 for image in raw_remote_images:
@@ -64,10 +73,14 @@ def get_image():
     #    new_image_list.append(f"./assets/{image}")
     big_image_list = local_images + remote_images
     image = random.choice(big_image_list)
-    if image[image.find("/")+1:] not in local_images:
-        remote_zip.extract(image, f"./assets/fetched")
-        os.rename(f"./assets/fetched/{image}", f"./assets/fetched/{image[image.find("/")+1:]}")
-    return send_file(f"./assets/fetched/{image[image.find("/")+1:]}", mimetype="image")
+    if not readonly_fs:
+        if image[image.find("/")+1:] not in local_images:
+            remote_zip.extract(image, f"./assets/fetched")
+            os.rename(f"./assets/fetched/{image}", f"./assets/fetched/{image[image.find("/")+1:]}")
+        return send_file(f"./assets/fetched/{image[image.find("/")+1:]}", mimetype="image")
+    else:
+        image_data = remote_zip.read(image)
+        return send_file(io.BytesIO(image_data), mimetype="image")
 
     return #send_file(io.BytesIO(requests.get(random.choice(local_images)).content), mimetype="image") # send_file(random.choice(new_image_list), mimetype="image")
 
@@ -93,7 +106,7 @@ def do_selenium_getting(): # Might not use because vercel can't use selenium
 
 
 
-app.run()
+#app.run()
 
 # regex for image links with w h at the end https:\/\/lh3\.googleusercontent\.com\/pw\/[a-zA-Z0-9-_]+=w[0-9]+-h[0-9]+
 # regex for all image links https:\/\/lh3\.googleusercontent\.com\/pw\/[a-zA-Z0-9-_]+
